Actualizacion de features
Solo queda conectar subir informe con back y poder ver los informes creados

+++ src/api/services/reportService.js
@@ -1,68 +1,119 @@
-import { ReportGINSHT, ReportPWD } from "../models/reportModel";
-import employeeService from "./employeeService";
-import companyService from "./companyService";
-import Company from "../models/companyModel";
+import ReportGINSHT from "../models/reportModel.js";
+import  ReportPWD from "../models/reportModel.js";
+import employeeService from "./employeeService.js";
+import companyService from "./companyService.js";
+import Company from "../models/companyModel.js";
+import HttpError from "../../utils/HttpError.js";
 
 const reportService = {
     getAllReports: async (userId) => {
+        if (!userId) {
+            throw new HttpError(400, 'Falta el userId');
+        }
         try {
             const reports = await ReportGINSHT.find({ User: userId }).exec();
+            if (!reports || reports.length === 0) {
+                throw new HttpError(404, 'No se encontraron informes para el usuario proporcionado.');
+            }
             reports.push(await ReportPWD.find({ User: userId }).exec());
             return reports;
         } catch (error) {
+            if (error instanceof HttpError) {
+                throw error;
+            }
             console.error('Error al buscar informes:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     getAllReportsEmployee: async (userId, employee) => {
-        try{
-            const reports = getAllReports(userId);
-            const employeeFound = await employeeService.findEmployee({ DNI: employee });
-            reports.filter(report => report.employee === employeeFound.DNI);
+        if (!userId) {
+            throw new HttpError(400, 'Falta el userId');
+        }
+        if (!employee) {
+            throw new HttpError(400, 'Falta el DNI del empleado');
+        }
+        try {
+            const reports = await getAllReports(userId);
+            if (!reports || reports.length === 0) {
+                throw new HttpError(404, 'No se encontraron informes para el usuario proporcionado.');
+            }
 
-            return reports;
+            const employeeFound = await employeeService.findEmployee({ DNI: employee });
+            if (!employeeFound) {
+                throw new HttpError(404, 'No se encontró el empleado con el DNI proporcionado.');
+            }
+
+            const filteredReports = reports.filter(report => report.employee === employeeFound.DNI);
+            if (!filteredReports || filteredReports.length === 0) {
+                throw new HttpError(204, 'No se encontraron informes para el empleado especificado.');
+            }
+
+            return filteredReports;
         } catch (error) {
+            if (error instanceof HttpError) {
+                throw error;
+            }
             console.error('Error al buscar informes:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     getAllReportsCompany: async (userId, company) => {
-        try{
-            const reports = getAllReports(userId);
-            const companyFound = await companyService.findCompany({ name: company });
-            reports.filter(report => report.company === companyFound.CIF);
+        if (!userId) {
+            throw new HttpError(400, 'Falta el userId');
+        }
+        if (!company) {
+            throw new HttpError(400, 'Falta el CIF de la empresa');
+        }
+        try {
+            const reports = await reportService.getAllReports(userId);
+            const companyFound = await companyService.findCompany({ CIF: company });
+            if (!companyFound) {
+                throw new HttpError(404, 'No se encontró la empresa con el CIF proporcionado.');
+            }
 
-            return reports;
+            const filteredReports = reports.filter(report => report.company.toString() === companyFound._id.toString());
+            if (filteredReports.length === 0) {
+                throw new HttpError(204, 'No se encontraron informes para la empresa especificada.');
+            }
+
+            return filteredReports;
         } catch (error) {
             console.error('Error al buscar informes:', error);
-            return null;
+            throw error; // Relanzamos el error
         }
     },
     getOneReport: async (userId, reportId) => {
+        if (!userId || !reportId) {
+            throw new HttpError(400, 'Falta el userId o el reportId');
+        }
         try {
-            // Buscar el informe por su ID en los dos modelos
             const reportGINSHT = await ReportGINSHT.findById(reportId).exec();
             const reportPWD = await ReportPWD.findById(reportId).exec();
-            // Si no se encuentra en ninguno, devolver null
             if (!reportGINSHT && !reportPWD) {
-                return null;
+                throw new HttpError(404, 'No se encontró el informe solicitado.');
             }
-            // Si se encuentra en alguno, devolver el informe encontrado
-            return reportGINSHT || reportPWD;
+            let report = reportGINSHT || reportPWD;
+            if (report.User.toString() !== userId) {
+                throw new HttpError(403, 'No está autorizado para ver este informe.');
+            }
+            return report;
         } catch (error) {
             console.error('Error al buscar informe por ID:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     createReportGINSHT: async (userId, reportData) => {
+        if (!userId || !reportData) {
+            throw new HttpError(400, 'Falta información necesaria para crear el informe.');
+        }
         try {
             const company = await companyService.findCompanyByCIF(reportData.company);
             if (!company) {
-                return null;
+                throw new HttpError(404, 'No se encontró la empresa especificada.');
             }
             const employee = await employeeService.findEmployee({ DNI: reportData.employee });
             if (!employee) {
-                return null;
+                throw new HttpError(404, 'No se encontró el empleado especificado.');
             }
             const report = new ReportGINSHT({
                 ...reportData,
@@ -74,18 +125,21 @@ const reportService = {
             return report;
         } catch (error) {
             console.error('Error al crear informe GINSHT:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     createReportPWD: async (userId, reportData) => {
+        if (!userId || !reportData) {
+            throw new HttpError(400, 'Falta información necesaria para crear el informe.');
+        }
         try {
             const company = await companyService.findCompanyByCIF(reportData.company);
             if (!company) {
-                return null;
+                throw new HttpError(404, 'No se encontró la empresa especificada.');
             }
             const employee = await employeeService.findEmployee({ DNI: reportData.employee });
             if (!employee) {
-                return null;
+                throw new HttpError(404, 'No se encontró el empleado especificado.');
             }
             const report = new ReportPWD({
                 ...reportData,
@@ -97,22 +151,26 @@ const reportService = {
             return report;
         } catch (error) {
             console.error('Error al crear informe PWD:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     modifyGINSHT: async (userId, reportId, reportData) => {
+        if (!userId || !reportId || !reportData) {
+            throw new HttpError(400, 'Falta información necesaria para modificar el informe.');
+        }
         try {
             const report = await ReportGINSHT.findById(reportId).exec();
             if (!report) {
-                return null;
+                throw new HttpError(404, 'No se encontró el informe especificado.');
             }
+            // Similar a create, verificamos la empresa y el empleado antes de proceder
             const company = await companyService.findCompanyByCIF(reportData.company);
             if (!company) {
-                return null;
+                throw new HttpError(404, 'No se encontró la empresa especificada.');
             }
             const employee = await employeeService.findEmployee({ DNI: reportData.employee });
             if (!employee) {
-                return null;
+                throw new HttpError(404, 'No se encontró el empleado especificado.');
             }
             report.set({
                 ...reportData,
@@ -123,22 +181,26 @@ const reportService = {
             return report;
         } catch (error) {
             console.error('Error al modificar informe GINSHT:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     modifyPWD: async (userId, reportId, reportData) => {
+        if (!userId || !reportId || !reportData) {
+            throw new HttpError(400, 'Falta información necesaria para modificar el informe.');
+        }
         try {
             const report = await ReportPWD.findById(reportId).exec();
             if (!report) {
-                return null;
+                throw new HttpError(404, 'No se encontró el informe especificado.');
             }
+            // Similar a create, verificamos la empresa y el empleado antes de proceder
             const company = await companyService.findCompanyByCIF(reportData.company);
             if (!company) {
-                return null;
+                throw new HttpError(404, 'No se encontró la empresa especificada.');
             }
             const employee = await employeeService.findEmployee({ DNI: reportData.employee });
             if (!employee) {
-                return null;
+                throw new HttpError(404, 'No se encontró el empleado especificado.');
             }
             report.set({
                 ...reportData,
@@ -149,31 +211,41 @@ const reportService = {
             return report;
         } catch (error) {
             console.error('Error al modificar informe PWD:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     deleteReportGINSHT: async (userId, reportId) => {
+        if (!userId || !reportId) {
+            throw new HttpError(400, 'Falta información necesaria para eliminar el informe.');
+        }
         try {
             const report = await ReportGINSHT.findByIdAndDelete(reportId).exec();
             if (!report) {
-                return null;
+                throw new HttpError(404, 'No se encontró el informe especificado para eliminar.');
             }
             return report;
-        } catch (error) {
+        } catch (
+        error) {
             console.error('Error al eliminar informe GINSHT:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
     deleteReportPWD: async (userId, reportId) => {
+        if (!userId || !reportId) {
+            throw new HttpError(400, 'Falta información necesaria para eliminar el informe.');
+        }
         try {
             const report = await ReportPWD.findByIdAndDelete(reportId).exec();
             if (!report) {
-                return null;
+                throw new HttpError(404, 'No se encontró el informe especificado para eliminar.');
             }
             return report;
-        } catch (error) {
+        } catch (
+        error) {
             console.error('Error al eliminar informe PWD:', error);
-            return null;
+            throw new HttpError(500, 'Error interno del servidor');
         }
     },
-}
+};
+
+export default reportService;
